Handle requests one at a time in an overwrite mode
With this we can now run multiple requests in a row. Whenever one of them fails the test will stop and the snapshot for that given response will be updated. We are also able to use variables throughout the entire sequence meaning that once they are defined they are available. This also includes storing fields that come from the previous response inside a variable